Redesign

+++ app/src/main/java/com/example/fakestore/model/network/NetworkRating.kt
@@ -2,5 +2,5 @@ package com.example.fakestore.model.network
 
 data class NetworkRating(
     val count: Int,
-    val rate: Double
+    val rate: Float
 )
